refactored Version info
