+++ crates/rbuilder/src/telemetry/servers/full.rs
@@ -7,17 +7,15 @@
 //!
 //! Full server may expose metrics that could leak information when running tdx.
 
+use rbuilder_utils::build_info::Version;
 use std::net::SocketAddr;
 use time::OffsetDateTime;
 use warp::{Filter, Rejection, Reply};
 
-use crate::{
-    telemetry::{
-        metrics::{gather_prometheus_metrics, set_version},
-        BUILDER_BALANCE, CURRENT_BLOCK, MAX_FRESH_GAUGE_AGE, ORDERPOOL_BUNDLES, ORDERPOOL_TXS,
-        ORDERPOOL_TXS_SIZE, REGISTRY,
-    },
-    utils::build_info::Version,
+use crate::telemetry::{
+    metrics::{gather_prometheus_metrics, set_version},
+    BUILDER_BALANCE, CURRENT_BLOCK, MAX_FRESH_GAUGE_AGE, ORDERPOOL_BUNDLES, ORDERPOOL_TXS,
+    ORDERPOOL_TXS_SIZE, REGISTRY,
 };
 
 pub async fn spawn(addr: SocketAddr, version: Version) -> eyre::Result<()> {
